Add typelen parameter to rfc1524_mailcap_lookup().
Because of mime_lookup commands, the call to mutt_check_lookup_list()
inside the function can modify the passed in type.  Add an explicit
length parameter to the function, rather than assume the parameter
size.  This also makes it more evident the type parameter can be
modified to callers.

Change the len parameter to mutt_check_lookup_list() to type size_t,
just to be correct about it.

+++ protos.h
@@ -236,7 +236,7 @@ const char *mutt_getcwd (BUFFER *);
 void mutt_help (int);
 const char *mutt_idxfmt_hook (const char *, CONTEXT *, HEADER *);
 void mutt_draw_tree (CONTEXT *);
-void mutt_check_lookup_list (BODY *, char *, int);
+void mutt_check_lookup_list (BODY *, char *, size_t);
 void mutt_make_attribution (CONTEXT *ctx, HEADER *cur, FILE *out);
 void mutt_make_forward_subject (ENVELOPE *env, CONTEXT *ctx, HEADER *cur);
 void mutt_make_help (char *, size_t, const char *, int, int);
